+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v 2.4.0 (?)
+# v 2.4.0 (2022-12-05)
 Changes in this release:
  - Update default Inmanta EULA header
  - Enhance module v2 generation support.